Tambah form options untuk closing dan opening

+++ OlahCitra.Core/ImageProcessing.cs
@@ -729,7 +729,7 @@ namespace OlahCitra.Core
             Bitmap outImage = new Bitmap(input.Width, input.Height);
 
             using (var image = input.ToImage<Bgr, byte>())
-            using (var imageHsv = image.Convert<Hsv, byte>())
+            using (var imageHsv = image.Convert<Hsv, int>())
             using (var mask = new UMat())
             using (var result = new UMat())
             {
